Some refactoring. Moved spectrum to its own file
